fix(daemon): isolate audit draw integration (#619)
